AuditorBundle/hl18/pyproject.toml
[project]
name = "hijaiyyahlang"
version = "0.1.0"
description = "Virtual HijaiyyahLang (HL-18) engine + release verifier"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
solver = ["ortools>=9.8.0"]
dev = ["pytest>=8.0.0", "ruff>=0.5.0"]

[project.scripts]
hl18 = "hijaiyyahlang.cli:main"
verify-hl18-release = "hijaiyyahlang.verify_release:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
